requirements.txt
Flask==3.1.0
marshmallow==4.0.0
flask-marshmallow==1.3.0
pandas==2.2.3
requests==2.32.3
flask-cors==6.0.0
Flask-RESTful==0.3.10
psycopg2==2.9.10
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
